Turn stub can_extract_subgraphs test into a snapshot test
