+++ tests/extract_subgraphs.rs
@@ -4,6 +4,107 @@ use apollo_compiler::Node;
 use apollo_federation::Supergraph;
 
 #[test]
+fn can_extract_subgraph() {
+    let schema = r#"
+      schema
+        @link(url: "https://specs.apollo.dev/link/v1.0")
+        @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
+      {
+        query: Query
+      }
+
+      directive @join__enumValue(graph: join__Graph!) repeatable on ENUM_VALUE
+
+      directive @join__field(graph: join__Graph, requires: join__FieldSet, provides: join__FieldSet, type: String, external: Boolean, override: String, usedOverridden: Boolean) repeatable on FIELD_DEFINITION | INPUT_FIELD_DEFINITION
+
+      directive @join__graph(name: String!, url: String!) on ENUM_VALUE
+
+      directive @join__implements(graph: join__Graph!, interface: String!) repeatable on OBJECT | INTERFACE
+
+      directive @join__type(graph: join__Graph!, key: join__FieldSet, extension: Boolean! = false, resolvable: Boolean! = true, isInterfaceObject: Boolean! = false) repeatable on OBJECT | INTERFACE | UNION | ENUM | INPUT_OBJECT | SCALAR
+
+      directive @join__unionMember(graph: join__Graph!, member: String!) repeatable on UNION
+
+      directive @link(url: String, as: String, for: link__Purpose, import: [link__Import]) repeatable on SCHEMA
+
+      enum E
+        @join__type(graph: SUBGRAPH2)
+      {
+        V1 @join__enumValue(graph: SUBGRAPH2)
+        V2 @join__enumValue(graph: SUBGRAPH2)
+      }
+
+      scalar join__FieldSet
+
+      enum join__Graph {
+        SUBGRAPH1 @join__graph(name: "Subgraph1", url: "https://Subgraph1")
+        SUBGRAPH2 @join__graph(name: "Subgraph2", url: "https://Subgraph2")
+      }
+
+      scalar link__Import
+
+      enum link__Purpose {
+        """
+        \`SECURITY\` features provide metadata necessary to securely resolve fields.
+        """
+        SECURITY
+
+        """
+        \`EXECUTION\` features provide metadata necessary for operation execution.
+        """
+        EXECUTION
+      }
+
+      type Query
+        @join__type(graph: SUBGRAPH1)
+        @join__type(graph: SUBGRAPH2)
+      {
+        t: T @join__field(graph: SUBGRAPH1)
+      }
+
+      type S
+        @join__type(graph: SUBGRAPH1)
+      {
+        x: Int
+      }
+
+      type T
+        @join__type(graph: SUBGRAPH1, key: "k")
+        @join__type(graph: SUBGRAPH2, key: "k")
+      {
+        k: ID
+        a: Int @join__field(graph: SUBGRAPH2)
+        b: String @join__field(graph: SUBGRAPH2)
+      }
+
+      union U
+        @join__type(graph: SUBGRAPH1)
+        @join__unionMember(graph: SUBGRAPH1, member: "S")
+        @join__unionMember(graph: SUBGRAPH1, member: "T")
+       = S | T
+    "#;
+
+    let supergraph = Supergraph::new(schema).unwrap();
+    let subgraphs = supergraph
+        .extract_subgraphs()
+        .expect("Should have been able to extract subgraphs");
+
+    let mut snapshot = String::new();
+    for (_name, subgraph) in subgraphs {
+        use std::fmt::Write;
+
+        _ = writeln!(
+            &mut snapshot,
+            "{}: {}\n---\n{}",
+            subgraph.name,
+            subgraph.url,
+            subgraph.schema.schema()
+        );
+    }
+    insta::assert_snapshot!(snapshot);
+}
+
+#[test]
 fn preserve_default_values_of_input_fields() {
     let supergraph = Supergraph::new(r#"
     schema
